+++ Controllers/StockContoller.cs
@@ -25,16 +25,24 @@ public class StockContoller : ControllerBase
 
     [HttpGet]
     public async Task<IActionResult> GetAll()
+
     {
+
+        if (!ModelState.IsValid) {
+            return BadRequest(ModelState);
+        }        
         var Stocks = await  _stockRepo.GetAllAsync();
         var stockDto =  Stocks.Select(s => s.ToStockDto());
         return Ok(stockDto);
 
     }
 
-    [HttpGet("{id}")]
+    [HttpGet("{id:int}")]
     public async Task<IActionResult> GetById([FromRoute] int id)
     {
+        if (!ModelState.IsValid) {
+            return BadRequest(ModelState);
+        }  
         var stock = await _stockRepo.GetByIdAsync(id);
         if (stock == null)
         {
@@ -48,6 +56,9 @@ public class StockContoller : ControllerBase
     [HttpPost]
     public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
     {
+          if (!ModelState.IsValid) {
+            return BadRequest(ModelState);
+        }  
 
         var stockModel = stockDto.ToStockFromCreateDTO();
         await _stockRepo.CreateAsync(stockModel);
@@ -55,10 +66,13 @@ public class StockContoller : ControllerBase
     }
 
 
-    [HttpPut("{id}")]
+    [HttpPut("{id:int}")]
 
     public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto)
     {
+        if (!ModelState.IsValid) {
+            return BadRequest(ModelState);
+        }  
 
         var stockModel = await _stockRepo.UpdateAsync(id , updateDto);
 
@@ -71,10 +85,13 @@ public class StockContoller : ControllerBase
         return Ok(stockModel.ToStockDto());
     }
 
-    [HttpDelete("{id}")]
+    [HttpDelete("{id:int}")]
 
     public async Task<IActionResult> Delete([FromRoute] int id)
     {
+        if (!ModelState.IsValid) {
+            return BadRequest(ModelState);
+        }  
         var stockModel = await _stockRepo.DeleteAsync(id);
 
         if (stockModel == null)
